requestparam changed to pathvariable in ratingcontroller

+++ service-rating/src/main/java/com/codecool/servicerating/entity/Rating.java
@@ -18,12 +18,12 @@ public class Rating {
 
     @Id
     @GeneratedValue
-    private Integer id;
+    private Long id;
 
     private String comment;
 
     private Integer rating;
 
-    private Integer waterId;
+    private Long waterId;
 
 }
